+++ crates/web-sys/src/features/mod.rs
@@ -6394,6 +6394,18 @@ mod gen_RtcRtcpParameters;
 #[cfg(feature = "RtcRtcpParameters")]
 pub use gen_RtcRtcpParameters::*;
 
+#[cfg(feature = "RtcRtpCapabilities")]
+#[allow(non_snake_case)]
+mod gen_RtcRtpCapabilities;
+#[cfg(feature = "RtcRtpCapabilities")]
+pub use gen_RtcRtpCapabilities::*;
+
+#[cfg(feature = "RtcRtpCodecCapability")]
+#[allow(non_snake_case)]
+mod gen_RtcRtpCodecCapability;
+#[cfg(feature = "RtcRtpCodecCapability")]
+pub use gen_RtcRtpCodecCapability::*;
+
 #[cfg(feature = "RtcRtpCodecParameters")]
 #[allow(non_snake_case)]
 mod gen_RtcRtpCodecParameters;
@@ -6411,6 +6423,12 @@ pub use gen_RtcRtpContributingSource::*;
 mod gen_RtcRtpEncodingParameters;
 #[cfg(feature = "RtcRtpEncodingParameters")]
 pub use gen_RtcRtpEncodingParameters::*;
+
+#[cfg(feature = "RtcRtpHeaderExtensionCapability")]
+#[allow(non_snake_case)]
+mod gen_RtcRtpHeaderExtensionCapability;
+#[cfg(feature = "RtcRtpHeaderExtensionCapability")]
+pub use gen_RtcRtpHeaderExtensionCapability::*;
 
 #[cfg(feature = "RtcRtpHeaderExtensionParameters")]
 #[allow(non_snake_case)]